feat: improve comment and create example
