+++ example/trace/provider/internal/consts.go
@@ -1,0 +1,17 @@
+// Copyright GoFrame Author(https://goframe.org). All Rights Reserved.
+//
+// This Source Code Form is subject to the terms of the MIT License.
+// If a copy of the MIT was not distributed with this file,
+// You can obtain one at https://github.com/gogf/gf.
+
+package internal
+
+const (
+	GRPCServiceName      = "otlp-grpc-client"
+	Endpoint             = "tracing-analysis-dc-bj.aliyuncs.com:8090"
+	TraceToken           = "******_******"
+	HTTPServiceName      = "otlp-http-client"
+	HTTPEndpoint         = "tracing-analysis-dc-hz.aliyuncs.com"
+	HTTPPath             = "adapt_******_******/api/otlp/traces"
+	TracerHostnameTagKey = "hostname"
+)
